Puck object slots are now permanently reserved for pucks and can not be reused by players later, this should fix some odd replay bugs

+++ src/hqm_server.rs
@@ -1916,7 +1916,7 @@ pub(crate) struct HQMServerConfiguration {
     pub(crate) time_intermission: u32,
     pub(crate) offside: HQMOffsideConfiguration,
     pub(crate) icing: HQMIcingConfiguration,
-    pub(crate) warmup_pucks: u32,
+    pub(crate) warmup_pucks: usize,
     pub(crate) limit_jump_speed: bool,
 
     pub(crate) cheats_enabled: bool,
